separate tests, update docstrings
